Refactoring #2 - Extract Method
After reading through my notes and the book i relies that the validID
method that i created just to make things easier is actually a major
refactor called Extract Method so I am adding this new Version to show
that this is a refactor.  Also I am still new to version control in
general and relies that when i committed the last revision that i forgot
to save all so cat.java and dog.java were never added to the repository
so that is why they are listed here.  In those files you can see the
getter and setters for the animal specific characteristics.  Sorry for
the confusion if any i will try to be more accurate with these commits.