Refactoring #3 - Substitute Algorithm
I started running into issues when i was trying to display all of the
animals after creating the super class.  I decided the best way to solve
this problem was to create an Array list of Animals instead of having
arrays of only cats and only dogs.   With this array list the code for
displaying all of the cats and dogs became easier to read.  I did have
to create and overloaded method in both Cats.java as well as Dogs.java
to allow for the printing of the animals.  This could not be in the
super class in this case do to the fact that the different animals do
have slightly different traits.  Also up until this point i had been
running manual tests which was in effecient so i started to add junit
testing.