Refactoring #4 - Duplicate Code and Documentation
I had commented out the method getAnimalInformation when i was doing the
third refactoring because i wanted to work on and test the first listing
method before working on the broken second one.  When i went back to
make it work i relised that there was no calling code this method at
all.  I loaded up the original source code, deleted this method and ran
the code to confirm that it was never used.  Since i was removing an
entire method anyways i decided to clean up the comments in the same
file since some of it was no longer relevant to the program.