Refactoring #5 - Rename Variable
My final refactoring in this assignment was to do a simple rename
variable. In the VetManager.java class i found it strange that the
variable vetclinic was used considering vetClinic was also a class
within this project.  For some reason that to me was a bad small.  I
also found that in the vetClinic class that there was a variable simply
called name.  I decided to change this to something more descriptive.  I
decided not to run any unit testing on this due to the fact that the
compiler would find name mismatches as its not a logic error.